// app.client/src/api/models/DocumentUrlListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DocumentUrlListVM = {
  documentID?: string | null;
  name?: string | null;
  description?: string | null;
  blobUrl?: string | null;
  category?: string | null;
  createdDate?: string;
  categoryID?: string;
  url?: string | null;
  extension?: string | null;
  documentType?: string | null;
  documentTypeId?: string | null;
};